+++ backend/canisters/local_user_index/api/src/lifecycle/init.rs
@@ -28,5 +28,7 @@ pub struct Args {
     pub media_scan_config: MediaScanConfig,
     #[serde(default)]
     pub multi_user_canisters_enabled: bool,
+    #[serde(default)]
+    pub call_push_enabled: bool,
     pub test_mode: bool,
 }